Translations with replaced keys did not trigger a MissingTranslationException
Ex:
Keys:
"strike_number_1" : "Strike One"

Key not translated and not flagged
$strikes = 1;
__('He makes contact with the ball, but not enough to get a hit. Foul tip. :strike_number.', [
  'strike_number' => __('strike_number_' + $strikes)
]);

The current check will compare
"He makes contact with the ball, but not enough to get a hit. Foul tip. :strike_number."
with
"He makes contact with the ball, but not enough to get a hit. Foul tip. Strike One"
and it will not flag the key as not translated.